requirements.txt
hrequests
beautifulsoup4
pyyaml
click
